Final touches - version & openapi
