+++ src/main/java/gov/usgs/wma/waterdata/controller/VersionController.java
@@ -1,5 +1,6 @@
 package gov.usgs.wma.waterdata.controller;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,9 +13,12 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 
-@Tag(name="Application Version", description="Data Download")
+@Tag(name="Application Version", description="Display")
 @RestController
 public class VersionController {
+
+	@Value("${site.url.base}")
+	private String serverUrl;
 
 	@Operation(
 			description="Return the web service version information.",
@@ -26,6 +30,6 @@ public class VersionController {
 			produces=MediaType.APPLICATION_JSON_VALUE
 	)
 	public RedirectView getVersion(RedirectAttributes attributes) {
-		return new RedirectView("/about/info", true, false);
+		return new RedirectView(serverUrl + "about/info", true, true);
 	}
 }
