--- src/main/java/gov/usgs/wma/waterdata/swagger/OpenApiConfig.java
@@ -1,21 +1,0 @@
-package gov.usgs.wma.waterdata.swagger;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-
-@Configuration
-public class OpenApiConfig {
-
-	@Bean
-	public OpenAPI customOpenAPI() {
-		return new OpenAPI()
-				.components(new Components())
-				.info(new Info().title("Contact Application API").description(
-						"This is a sample Spring Boot RESTful service using springdoc-openapi and OpenAPI 3."));
-
-	}
-}
